Split Carousel into multiple files
