+++ client/src/components/shared/Carousel/Carousel.css
@@ -26,49 +26,3 @@
 
   overflow: hidden;
 }
-
-.carousel-content {
-  flex-grow: 1;
-}
-
-.carousel-content-link {
-  margin-left: 10px;
-}
-
-.carousel-content-author {
-  text-align: end;
-  font-style: italic;
-}
-
-.carousel-page-indicator {
-  margin-top: 10px;
-
-  display: flex;
-  flex-grow: 0;
-}
-
-.carousel-page-bar {
-  display: inline-block;
-
-  width: 6px;
-  height: 6px;
-
-  margin: 0 2px;
-
-  flex-grow: 1;
-
-  border-radius: 3px;
-  background-color: #bfccd6;
-}
-
-.carousel-page-bar:hover {
-  cursor: pointer;
-}
-
-.carousel-page-bar.active {
-  background-color: #5c7080;
-}
-
-.carousel-nav-button {
-  flex-grow: 0;
-}
