+++ client/src/components/shared/Carousel/CarouselPageIndicator/CarouselPageIndicator.css
@@ -1,0 +1,28 @@
+.carousel-page-indicator {
+  margin-top: 10px;
+
+  display: flex;
+  flex-grow: 0;
+}
+
+.carousel-page-bar {
+  display: inline-block;
+
+  width: 6px;
+  height: 6px;
+
+  margin: 0 2px;
+
+  flex-grow: 1;
+
+  border-radius: 3px;
+  background-color: #bfccd6;
+}
+
+.carousel-page-bar:hover {
+  cursor: pointer;
+}
+
+.carousel-page-bar.active {
+  background-color: #5c7080;
+}
